[mm] hugetlb: fix type of hugetlb_treat_as_movable variable
Message-id: <20170308160531.26797-6-aarcange@redhat.com>
Patchwork-id: 172898
O-Subject: [RHEL7.4 PATCH 05/59] mm: hugetlb: fix type of hugetlb_treat_as_movable variable
Bugzilla: 1430172
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>

From: Andrey Ryabinin <a.ryabinin@samsung.com>

hugetlb_treat_as_movable declared as unsigned long, but
proc_dointvec() used for parsing it:

static struct ctl_table vm_table[] = {
...
        {
                .procname	= "hugepages_treat_as_movable",
                .data		= &hugepages_treat_as_movable,
                .maxlen		= sizeof(int),
                .mode		= 0644,
                .proc_handler	= proc_dointvec,
        },

This seems harmless, but it's better to use int type here.

Cc: Dmitry Vyukov <dvyukov@google.com>
Cc: Manfred Spraul <manfred@colorfullife.com>
Acked-by: David Rientjes <rientjes@google.com>
(cherry picked from commit 753162cd849c45580fb5aaa7f3597c81e74e391c)

+++ include/linux/hugetlb.h
@@ -88,7 +88,7 @@ void putback_active_hugepage(struct page *page);
 pte_t *huge_pmd_share(struct mm_struct *mm, unsigned long addr, pud_t *pud);
 #endif
 
-extern unsigned long hugepages_treat_as_movable;
+extern int hugepages_treat_as_movable;
 extern const unsigned long hugetlb_zero, hugetlb_infinity;
 extern int sysctl_hugetlb_shm_group;
 extern struct list_head huge_boot_pages;
